feat(MiniApp(Company)): Make company mini-app
