+++ src/interfaces/http/modules/company/index.js
@@ -1,209 +1,13 @@
-const { Router } = require('express')
-const Status = require('http-status')
-const container = require('src/container') // we have to get the DI
-const { get, post, put, remove } = require('src/app/company')
+const container = require('src/container')
+const router = require('./router')
+const instance = require('./instance')
 
 module.exports = () => {
-  const router = Router()
-  const { repository: {
-    companyRepository
-  }, logger, auth, response: { Success, Fail } } = container.cradle
+  const { logger, response: { Success, Fail }, auth } = container.cradle
+  const app = instance()
 
-  const getUseCase = get({ companyRepository })
-  const postUseCase = post({ companyRepository })
-  const putUseCase = put({ companyRepository })
-  const deleteUseCase = remove({ companyRepository })
-
-  /**
- * @swagger
- * definitions:
- *   company:
- *     properties:
- *       id:
- *         type: string
- *         format: uuid
- *       name:
- *         type: string
- *       address:
- *         type: string
- *       contact:
- *         type: string
- *       tin:
- *         type: string
- *       sss:
- *         type: string
- *       philhealth:
- *         type: string
- *       isDeleted:
- *         type: number
- *       createdBy:
- *         type: string
- *         format: uuid
- */
-
-  router.use(auth.authenticate())
-
-  /**
-  * @swagger
-  * /companies:
-  *   get:
-  *     tags:
-  *       - Companies
-  *     description: Returns a list of companies
-  *     security:
-  *       - JWT: []
-  *     responses:
-  *       200:
-  *         description: An array of companies
-  *         schema:
-  *           type: array
-  *           items:
-  *             $ref: '#/definitions/company'
-  *       401:
-  *        $ref: '#/responses/Unauthorized'
-  */
-  router
-    .get('/', (req, res) => {
-      getUseCase
-        .all(req, res)
-        .then(data => {
-          res.status(Status.OK).json(Success(data))
-        })
-        .catch((error) => {
-          logger.error(error) // we still need to log every error for debugging
-          res.status(Status.BAD_REQUEST).json(
-            Fail(error.message))
-        })
-    })
-
-  /**
- * @swagger
- * /companies:
- *   post:
- *     tags:
- *       - Companies
- *     description: Create new company
- *     security:
- *       - JWT: []
- *     produces:
- *       - application/json
- *     parameters:
- *       - name: body
- *         description: Company's Entity
- *         in: body
- *         required: true
- *         type: string
- *         schema:
- *           $ref: '#/definitions/company'
- *     responses:
- *       200:
- *         description: Successfully Created
- *         schema:
- *           $ref: '#/definitions/company'
- *       401:
- *         $ref: '#/responses/Unauthorized'
- *       400:
- *         $ref: '#/responses/BadRequest'
- */
-  router
-    .post('/', (req, res) => {
-      postUseCase
-        .create({ body: req.body })
-        .then(data => {
-          res.status(Status.OK).json(Success(data))
-        })
-        .catch((error) => {
-          logger.error(error) // we still need to log every error for debugging
-          res.status(Status.BAD_REQUEST).json(
-            Fail(error.message))
-        })
-    })
-
-  /**
-   * @swagger
-   * /companies:
-   *   put:
-   *     tags:
-   *       - Companies
-   *     description: Update Company
-   *     security:
-   *       - JWT: []
-   *     produces:
-   *       - application/json
-   *     parameters:
-   *       - name: id
-   *         in: path
-   *         required: true
-   *         description: Company's ID to update
-   *         type: string
-   *       - name: body
-   *         description: Company's Entity
-   *         in: body
-   *         required: true
-   *         type: string
-   *         schema:
-   *           $ref: '#/definitions/company'
-   *     responses:
-   *       200:
-   *         description: Successfully Updated
-   *         schema:
-   *           $ref: '#/definitions/company'
-   *       401:
-   *         $ref: '#/responses/Unauthorized'
-   *       400:
-   *         $ref: '#/responses/BadRequest'
-   */
-  router
-    .put('/:id', (req, res) => {
-      putUseCase
-        .update({ id: req.params.id, body: req.body })
-        .then(data => {
-          res.status(Status.OK).json(Success(data))
-        })
-        .catch((error) => {
-          logger.error(error) // we still need to log every error for debugging
-          res.status(Status.BAD_REQUEST).json(
-            Fail(error.message))
-        })
-    })
-
-  /**
-   * @swagger
-   * /companies:
-   *   delete:
-   *     tags:
-   *       - Companies
-   *     description: Delete Company
-   *     security:
-   *       - JWT: []
-   *     produces:
-   *       - application/json
-   *     parameters:
-   *       - name: id
-   *         in: path
-   *         required: true
-   *         description: Company's ID to delete
-   *         type: string
-   *     responses:
-   *       200:
-   *         description: Successfully Deleted
-   *         schema:
-   *           $ref: '#/definitions/company'
-   *       401:
-   *         $ref: '#/responses/Unauthorized'
-   */
-  router
-    .delete('/:id', (req, res) => {
-      deleteUseCase
-        .remove({ id: req.params.id })
-        .then(data => {
-          res.status(Status.OK).json(Success(data))
-        })
-        .catch((error) => {
-          logger.error(error) // we still need to log every error for debugging
-          res.status(Status.BAD_REQUEST).json(
-            Fail(error.message))
-        })
-    })
-  return router
+  return {
+    app,
+    router: router({ logger, auth, response: { Success, Fail }, ...app })
+  }
 }
